move bp file export and import into 'bp/'
